# api/openapi/enhanced-business-intelligence-api.yaml
openapi: 3.0.3
info:
  title: Enhanced Business Intelligence System API
  description: |
    Enhanced Business Intelligence System API with intelligent routing, advanced classification,
    and comprehensive business intelligence features.
    
    ## Key Features
    - **Intelligent Routing**: Automatically routes requests to the best classification method
    - **Enhanced Classification**: Advanced business classification with 10+ data points
    - **Business Intelligence**: Analytics, insights, and trend analysis
    - **Backward Compatibility**: Full v1 API compatibility with enhanced functionality
    
    ## Migration Guide
    - **v1 endpoints**: Still supported but route through intelligent routing system
    - **v2 endpoints**: New enhanced endpoints with advanced features
    - **Benefits**: Improved performance, better accuracy, comprehensive data extraction
    
  version: 2.0.0
  contact:
    name: API Support
    email: api-support@business-verification.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.business-verification.com/v2
    description: Production server
  - url: https://staging-api.business-verification.com/v2
    description: Staging server
  - url: http://localhost:8080/v2
    description: Development server
...
paths:
...
  /classify:
    post:
      tags:
        - Classification
      summary: Enhanced single business classification
      description: |
        Perform enhanced business classification using intelligent routing system.
        This endpoint automatically selects the best classification method based on
        the business data and provides comprehensive results with 10+ data points.
      operationId: classifyBusiness
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedClassificationRequest'
            examples:
              basic_classification:
                summary: Basic business classification
                value:
                  business_name: "Acme Corporation"
                  website_url: "https://acme.com"
                  description: "Technology consulting services"
              comprehensive_classification:
                summary: Comprehensive business classification
                value:
                  business_name: "TechCorp Solutions"
                  website_url: "https://techcorp.com"
                  description: "Software development and consulting"
                  industry: "Technology"
                  keywords: "software, development, consulting"
                  geographic_region: "California, USA"
      responses:
        '200':
          description: Classification completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedClassificationResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key for authentication
  schemas:
    EnhancedClassificationRequest:
      type: object
      required:
        - business_name
      properties:
        business_name:
          type: string
          description: Name of the business to classify
          example: "Acme Corporation"
          minLength: 1
          maxLength: 255
        website_url:
          type: string
          description: Business website URL
          example: "https://acme.com"
          format: uri
        description:
          type: string
          description: Business description
          example: "Technology consulting services"
          maxLength: 1000
        industry:
          type: string
          description: Known industry or sector
          example: "Technology"
          maxLength: 100
        keywords:
          type: string
          description: Keywords related to the business
          example: "software, development, consulting"
          maxLength: 500
        geographic_region:
          type: string
          description: Geographic region or location
          example: "California, USA"
          maxLength: 100
        enhanced_features:
          type: object
          description: Enable specific enhanced features
          properties:
            include_company_size:
              type: boolean
              description: Include company size analysis
              default: true
            include_business_model:
              type: boolean
              description: Include business model analysis
              default: true
            include_technology_stack:
              type: boolean
              description: Include technology stack analysis
              default: true
            include_risk_assessment:
              type: boolean
              description: Include risk assessment
              default: false
    EnhancedClassificationResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique classification ID
          example: "class_1234567890"
        business_name:
          type: string
          description: Business name
          example: "Acme Corporation"
        status:
          type: string
          enum: [pending, completed, failed, error]
          description: Classification status
          example: "completed"
        classifications:
          type: array
          items:
            $ref: '#/components/schemas/IndustryClassification'
          description: Industry classifications
        company_size:
          $ref: '#/components/schemas/CompanySize'
        business_model:
          $ref: '#/components/schemas/BusinessModel'
        technology_stack:
          $ref: '#/components/schemas/TechnologyStack'
        risk_assessment:
          $ref: '#/components/schemas/RiskAssessment'
        metadata:
          $ref: '#/components/schemas/ClassificationMetadata'
        created_at:
          type: string
          format: date-time
          description: Classification creation timestamp
          example: "2024-12-19T10:30:00Z"
    IndustryClassification:
      type: object
      properties:
        industry_code:
          type: string
          description: Industry classification code
          example: "541511"
        industry_name:
          type: string
          description: Industry name
          example: "Custom Computer Programming Services"
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for this classification
          example: 0.95
        classification_type:
          type: string
          enum: [NAICS, SIC, MCC]
          description: Type of industry classification
          example: "NAICS"
    CompanySize:
      type: object
      properties:
        employee_count_range:
          type: string
          enum: [1-10, 11-50, 51-200, 201-500, 501-1000, 1000+]
          description: Estimated employee count range
          example: "51-200"
        revenue_indicator:
          type: string
          enum: [startup, small_business, medium_business, large_business]
          description: Revenue size indicator
          example: "medium_business"
        office_locations_count:
          type: integer
          minimum: 0
          description: Number of office locations
          example: 3
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for size estimation
          example: 0.85
    BusinessModel:
      type: object
      properties:
        model_type:
          type: string
          enum: [B2B, B2C, B2B2C, Marketplace, SaaS, Freemium, Subscription, OneTime]
          description: Business model type
          example: "B2B"
        revenue_model:
          type: string
          description: Primary revenue model
          example: "subscription"
        target_market:
          type: string
          description: Target market description
          example: "Enterprise technology companies"
        pricing_model:
          type: string
          description: Pricing model description
          example: "Tiered subscription pricing"
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for business model analysis
          example: 0.90
    TechnologyStack:
      type: object
      properties:
        programming_languages:
          type: array
          items:
            type: string
          description: Programming languages used
          example: ["JavaScript", "Python", "Go"]
        frameworks:
          type: array
          items:
            type: string
          description: Frameworks and libraries used
          example: ["React", "Node.js", "Django"]
        cloud_platforms:
          type: array
          items:
            type: string
          description: Cloud platforms used
          example: ["AWS", "Google Cloud"]
        third_party_services:
          type: array
          items:
            type: string
          description: Third-party services and integrations
          example: ["Stripe", "SendGrid", "MongoDB"]
        development_tools:
          type: array
          items:
            type: string
          description: Development tools and platforms
          example: ["GitHub", "Docker", "Kubernetes"]
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for technology stack analysis
          example: 0.88
    RiskAssessment:
      type: object
      properties:
        overall_risk:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          description: Overall risk level
          example: "LOW"
        security_risk:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          description: Security risk level
          example: "LOW"
        financial_risk:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          description: Financial risk level
          example: "MEDIUM"
        compliance_risk:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          description: Compliance risk level
          example: "LOW"
        risk_factors:
          type: array
          items:
            type: string
          description: Identified risk factors
          example: ["Limited financial history", "New market entry"]
    ClassificationMetadata:
      type: object
      properties:
        processing_time:
          type: string
          description: Total processing time
          example: "1.2s"
        strategies_used:
          type: array
          items:
            type: string
          description: Classification strategies used
          example: ["intelligent_routing", "ml_classification", "keyword_analysis"]
        cache_hit:
          type: boolean
          description: Whether result was served from cache
          example: false
        intelligent_routing_method:
          type: string
          description: Method selected by intelligent routing
          example: "ml_classification"
        data_points_extracted:
          type: integer
          description: Number of data points extracted
          example: 12
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Error message
              example: "Invalid business name provided"
            details:
              type: object
              description: Additional error details
            request_id:
              type: string
              description: Request ID for tracking
              example: "req_1234567890"
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
              example: "2024-12-19T10:30:00Z"